Convert JumpTargetInstructionWrapper to VirtualLabel
Wrapping instructions that were targets of jumps with another
IInstruction was obscuring the interfaces of the wrapped
instruction which was breaking code that relied on being able
to access them. It seemed simpler to just convert
JumpTargetInstructionWrapper to its own standalone IInstruction
than try to expose the interfaces of the wrapped instruction.